app/model_service/requirements.txt
absl-py==1.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chex==0.1.3
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
dalle-mini==0.1.1
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.7
docker-pycreds==0.4.0
einops==0.4.1
emoji==2.0.0
entrypoints==0.4
etils==0.6.0
fastjsonschema==2.16.1
filelock==3.7.1
flax==0.5.3
fonttools==4.34.4
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jax==0.3.15
jaxlib==0.3.15
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.21.6
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
rich==11.2.0
scipy==1.7.3
Send2Trash==1.8.0
sentry-sdk==1.8.0
setproctitle==1.3.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
tensorstore==0.1.21
terminado==0.15.0
tinycss2==1.1.1
tokenizers==0.12.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.11
uvicorn==0.18.2
vqgan-jax @ git+https://github.com/patil-suraj/vqgan-jax.git@10ef240f8ace869e437f3c32d14898f61512db12
wandb==0.12.21
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zipp==3.8.1
